routellm.net/src/RouteLLM.Routers/MatrixFactorizationRouter.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Numerics;
using RouteLLM.Core;

namespace RouteLLM.Routers
{
    public class MatrixFactorizationRouter : Router
    {
        private readonly int embeddingSize;
        private readonly int numModels;
        private Vector<float>[] modelEmbeddings;
        private Vector<float>[] promptEmbeddings;

        public MatrixFactorizationRouter(int embeddingSize, int numModels)
        {
            this.embeddingSize = embeddingSize;
            this.numModels = numModels;
            InitializeEmbeddings();
        }

        private void InitializeEmbeddings()
        {
            var random = new Random();
            modelEmbeddings = Enumerable.Range(0, numModels)
                .Select(_ => Vector<float>.Build.Dense(embeddingSize, i => (float)random.NextDouble()))
                .ToArray();
            promptEmbeddings = new Vector<float>[0]; // Initialize as empty, will be populated later
        }

        public override ModelPair Route(string prompt)
        {
            var promptEmbedding = GetPromptEmbedding(prompt);
            var scores = modelEmbeddings.Select(m => Vector.Dot(promptEmbedding, m)).ToArray();
            var bestModelIndex = Array.IndexOf(scores, scores.Max());
            
            // Assuming models are numbered from 0 to numModels-1
            return new ModelPair(bestModelIndex.ToString(), (bestModelIndex + 1) % numModels.ToString());
        }

        private Vector<float> GetPromptEmbedding(string prompt)
        {
            // TODO: Implement actual embedding logic
            // For now, return a random vector
            var random = new Random();
            return Vector<float>.Build.Dense(embeddingSize, i => (float)random.NextDouble());
        }

        // TODO: Implement training method
        public void Train(List<string> prompts, List<int> labels)
        {
            // Implementation for training the model
        }
    }
}
